Initial Commit
Code can currently compute the degree score for any year for any list of modules with any composition of assessments.

Currently need to write checks that the assessment weights of a module total to 100% or less, if it is less, give a warning that the modlue is partially complete.

Also could implement features to print an academic transcript, interface with a database, analytics on assessment types etc.