Added Target Score Calculator
Wrote a method to the UniversityDegree class, which given your current degree, allows you to see what average you require
for the remainder of your degree to obtain a given grade, so long as this required average is not greater than 100.

This function works for any stage in a degree, even when part way through a year.